fix: multiple bug fixes for arena coach + matchmaker stability
- Fix in_progress_at set to wrong value (assignment ID instead of
  timestamp), causing stuck games to never be cleaned up.
- Fix declined assignments never retrying — now set to 'retry' status
  with escalating delay (30s→480s, max 5 retries).
- Add GTP health check before coach reports 'ready', catching engines
  that crash on startup (like edax missing eval.dat).
- Fix Elo uncertainty using wrong engine version in matchmaker
  (GetEngineIDByName → GetEngineID with exact name+version).
- Add 90s timeout for 'accepted' assignments.
- Deferred relay session cleanup on match failure — fixes engines
  sitting idle consuming cores for 10 minutes after a failed match.
- Fix opening move race: send opening moves synchronously instead of
  fire-and-forget + drain, preventing stale responses from leaking
  into the genmove phase and being misinterpreted as moves.
- Add %game_time% substitution in coach (player YAML args). Lets
  engines declare their time flag convention (e.g. -t %game_time%).
- Add arena-clear-db.sh script to clear game data while keeping
  api_tokens and web_sessions.
- Update Documentation for %game_time% substitution and arena-clear-db.sh.

+++ internal/web/web.go
@@ -216,7 +216,7 @@ func (h *Handler) renderEloChart(w http.ResponseWriter, r *http.Request) {
 		fmt.Fprintf(w, `<line x1="0" y1="%.0f" x2="%.0f" y2="%.0f" stroke="var(--border)" stroke-width="0.5"/>`, y, graphw, y)
 		fmt.Fprintf(w, `<text x="-6" y="%.0f" text-anchor="end" fill="var(--muted)" font-size="10">%.0f</text>`, y+4, val)
 	}
-	fmt.Fprintf(w, `<text x="%.0f" y="%.0f" text-anchor="middle" fill="var(--fg)" font-size="12">Elo</text>`, graphw/2, -25)
+	fmt.Fprintf(w, `<text x="%.0f" y="%.0f" text-anchor="middle" fill="var(--fg)" font-size="12">Elo</text>`, float64(graphw)/2, float64(-25))
 	fmt.Fprintf(w, `<text x="%.0f" y="%.0f" text-anchor="middle" fill="var(--muted)" font-size="10">Time →</text>`, graphw/2, graphh+30)
 	// Data lines
 	for i, data := range engineData {
